linux.palette: update foreground/background
The Linux console uses 0xAAAAAA for the foreground on (black) background
so update this to match.

There isn't a precident for light mode so try to keep the same distance in
contrast with 0x555555 (+ 0xAAAAAA = 0xFFFFFF).

We may need to tweak the palette for light, but lets start with this and
see how things go from there.

Fixes: #396