window: avoid saving session twice during shutdown
This fixes an issue where we would save the session before shutting down
and then save the session incrementally due to pages getting removed as
part of the "ensure tab is destroyed" sequence.

This avoids those additional session saves so we keep the session we
really cared about (e.g. before the close window procedure).

Fixes: #312