window: use popover to filter new-terminal menu items
Instead of using GtkPopoverMenu, this implements a custom GtkPopover that
is somewhat similar to GtkDropDown. It allows filtering between the various
types of entries (container, local session, profiles) and tries to sort
them somewhat usefully.

The position of My Computer does change, primarily to make the section
headers look better as we cannot insert "separators" cleanly.

Styling tries to match things to menus but there could certainly be more
things we want to do there to make it feel nicer.

Searching is provided using GtkStringFilter, so all the caveats there
apply with what you can type to filter.

Fixes: #178