gnome.palette: Adopt kgx colors for dark scheme
This seems to be the color scheme that the design team agreed on back
in the day, so let's switch to that.

We're not adapting the light theme from kgx because they just reuse
the dark theme in light mode (and making an own light theme is a TODO
in kgx), so let's not touch the existing theme there.

Fixes: #437