app: defer cleanup to finalize
We don't really want this to happen at shutdown because some final main
loop flushing could allow for re-entry to call into API expecting these
objects to still exist.

Just defer to finalize. Realistically that won't happen either, but that
isn't an issue since we are about to exit anyway.